minor fix in statistic procesing

+++ src/main/java/sk/styk/martin/bakalarka/statistics/processors/DefinedPermissionsStatisticsProcessor.java
@@ -67,12 +67,12 @@ public class DefinedPermissionsStatisticsProcessor extends TopListProcessorBase 
                 List<String> protectonLevelList = manifestData.getPermissionsProtectionLevel();
 
                 if (permissionsList != null) {
-                    int listSize = permissionsList.size();
+                    Integer listSize = permissionsList.size();
                     permissionsNumbersList.add(new Double(listSize));
                     if (listSize != 0) {
                         permissionsNumbersListNonZero.add(new Double(listSize));
                     }
-                    permsRecordPair= processMaxExtreme("perms",listSize,data.getFileName());
+                    permsRecordPair = processMaxExtreme("perms", listSize.longValue(), data.getFileName());
 
                     for (String perm : protectonLevelList) {
                         if (topProtectionLevel.containsKey(perm)) {
